[mod] implement searx.wikidata_units for unit converters
